`torch_dtype` is deprecated! Use `dtype` instead!

+++ src/eval_trained_student.py
@@ -74,7 +74,7 @@ def main():
     if tokenizer.pad_token is None:
         tokenizer.pad_token = tokenizer.eos_token
 
-    model = AutoModelForCausalLM.from_pretrained(ckpt, torch_dtype=dtype)
+    model = AutoModelForCausalLM.from_pretrained(ckpt, dtype=dtype)
     model.to(device)
     model.eval()
